test_router_handover_leak: add new test for ZMQ_ROUTER_HANDOVER flag
This one tests that the issue #3238 won't resurrect.  The idea of the test
is simple: create router with handover and start two dealers in loop
with the same identity.  On each iteration router "rolls over" from one
dealer to another.  No socket leaks should happen.  We expect 5000
iterations should be enough not to crash the test.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>